Belgium Class for downloading, cleaning and processing notification data

Information for downloading, cleaning and processing COVID-19 region level 1 and 2 data for Belgium.

Source

https://epistat.sciensano.be/Data/COVID19BE_CASES_AGESEX.csv

https://epistat.sciensano.be/Data/COVID19BE_CASES_AGESEX.csv

https://epistat.sciensano.be/Data/COVID19BE_CASES_AGESEX.csv

Super class

covidregionaldata::DataClass -> Belgium

Public fields

origin

name of origin to fetch data for

supported_levels

A list of supported levels.

supported_region_names

A list of region names in order of level.

supported_region_codes

A list of region codes in order of level. ISO 3166-2 codes are used for both region and province levels in Belgium, and for provinces these are marked as being iso_3166_2_province

common_data_urls

List of named links to raw data that are common across levels.

level_data_urls

List of named links to raw data specific to each level of regions. For Belgium, there are only additional data for level 1 regions.

source_data_cols

existing columns within the raw data

Methods

Public methods

Belgium$set_region_codes()

Belgium$download()

Belgium$clean_level_1()

Belgium$clean_level_2()

Belgium$clone()

Inherited methods

covidregionaldata::DataClass$available_regions()

covidregionaldata::DataClass$clean()

covidregionaldata::DataClass$clean_common()

covidregionaldata::DataClass$filter()

covidregionaldata::DataClass$get()

covidregionaldata::DataClass$initialize()

covidregionaldata::DataClass$process()

covidregionaldata::DataClass$return()

covidregionaldata::DataClass$summary()

Method set_region_codes()

Set up a table of region codes for clean data

Usage

Belgium$set_region_codes()

Method set_region_codes()

Set up a table of region codes for clean data

Usage

Belgium$set_region_codes()

Method download()

Downloads data from source and (for Belgium) applies an initial data patch.

Usage

Belgium$download()

Method download()

Downloads data from source and (for Belgium) applies an initial data patch.

Usage

Belgium$download()

Method clean_level_1()

Region-level Data Cleaning

Usage

Belgium$clean_level_1()

Method clean_level_1()

Region-level Data Cleaning

Usage

Belgium$clean_level_1()

Method clean_level_2()

Province-level Data Cleaning

Usage

Belgium$clean_level_2()

Method clean_level_2()

Province-level Data Cleaning

Usage

Belgium$clean_level_2()

Method clone()

The objects of this class are cloneable with this method.

Usage

Belgium$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Method clone()

The objects of this class are cloneable with this method.

Usage

Belgium$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

if (FALSE) {
region <- Belgium$new(verbose = TRUE, steps = TRUE, get = TRUE, level = "2")
region$return()
}