Brazil Class for downloading, cleaning and processing notification data

Information for downloading, cleaning and processing COVID-19 region data for Brazil.

Data available on Github, curated by Wesley Cota: DOI 10.1590/SciELOPreprints.362

Source

https://github.com/wcota/covid19br

Super class

covidregionaldata::DataClass -> Brazil

Public fields

country

name of country to fetch data for

supported_levels

A list of supported levels.

supported_region_names

A list of region names in order of level.

supported_region_codes

A list of region codes in order of level.

common_data_urls

List of named links to raw data. Data is available at the city level and is aggregated to provide state data.

source_data_cols

existing columns within the raw data

Methods

Public methods

Brazil$set_region_codes()

Brazil$clean_common()

Brazil$clean_level_1()

Brazil$clean_level_2()

Brazil$clone()

Inherited methods

covidregionaldata::DataClass$clean()

covidregionaldata::DataClass$download()

covidregionaldata::DataClass$filter()

covidregionaldata::DataClass$get()

covidregionaldata::DataClass$initialize()

covidregionaldata::DataClass$process()

covidregionaldata::DataClass$return()

covidregionaldata::DataClass$summary()

Method set_region_codes()

Set up a table of region codes for clean data

Usage

Brazil$set_region_codes()

Method set_region_codes()

Set up a table of region codes for clean data

Usage

Brazil$set_region_codes()

Method clean_common()

Common data cleaning for both levels

Usage

Brazil$clean_common()

Method clean_common()

Common data cleaning for both levels

Usage

Brazil$clean_common()

Method clean_level_1()

State Level Data Cleaning

Usage

Brazil$clean_level_1()

Method clean_level_1()

State Level Data Cleaning

Usage

Brazil$clean_level_1()

Method clean_level_2()

City Level Data Cleaning

Usage

Brazil$clean_level_2()

Method clean_level_2()

City Level Data Cleaning

Usage

Brazil$clean_level_2()

Method clone()

The objects of this class are cloneable with this method.

Usage

Brazil$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Method clone()

The objects of this class are cloneable with this method.

Usage

Brazil$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

if (FALSE) {
region <- Brazil$new(verbose = TRUE, steps = TRUE, get = TRUE)
region$return()
}