Italy Class for downloading, cleaning and processing notification data

Information for downloading, cleaning and processing COVID-19 region data for Italy.

Source

https://github.com/pcm-dpc/COVID-19/

See also

Super class

covidregionaldata::DataClass -> Italy

Public fields

origin

name of origin to fetch data for

supported_levels

A list of supported levels.

supported_region_names

A list of region names in order of level.

supported_region_codes

A list of region codes in order of level.

common_data_urls

List of named links to raw data. The first, and only entry, is be named main.

source_data_cols

existing columns within the raw data

source_text

Plain text description of the source of the data

source_url

Website address for explanation/introduction of the data

Methods

Public methods

Italy$set_region_codes()

Italy$clean_common()

Italy$clone()

Inherited methods

covidregionaldata::DataClass$available_regions()

covidregionaldata::DataClass$clean()

covidregionaldata::DataClass$download()

covidregionaldata::DataClass$download_JSON()

covidregionaldata::DataClass$filter()

covidregionaldata::DataClass$get()

covidregionaldata::DataClass$initialize()

covidregionaldata::DataClass$process()

covidregionaldata::DataClass$return()

covidregionaldata::DataClass$summary()

covidregionaldata::DataClass$test()

Method set_region_codes()

Set up a table of region codes for clean data

Usage

Italy$set_region_codes()

Method clean_common()

State level data cleaning

Usage

Italy$clean_common()

Method clone()

The objects of this class are cloneable with this method.

Usage

Italy$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

if (FALSE) {
region <- Italy$new(verbose = TRUE, steps = TRUE, get = TRUE)
region$return()
}