Meixco Class for downloading, cleaning and processing notification data

Information for downloading, cleaning and processing COVID-19 region data for Mexico.

Notes on region codes:

Level 1 codes = ISO-3166-2, source: https://en.wikipedia.org/wiki/ISO_3166-2:MX

Level 2 codes = INEGI Mexican official statistics geocoding, source: raw data

Level 1 INEGI codes are the first 2 characters of Level 2 INEGI codes

Source

https://datos.covid-19.conacyt.mx/

See also

Super class

covidregionaldata::DataClass -> Mexico

Public fields

origin

name of origin to fetch data for

supported_levels

A list of supported levels.

supported_region_names

A list of region names in order of level.

supported_region_codes

A list of region codes in order of level.

common_data_urls

List of named links to raw data.

level_data_urls

List of named links to raw data that are level specific.

source_data_cols

existing columns within the raw data

source_text

Plain text description of the source of the data

source_url

Website address for explanation/introduction of the data

Methods

Public methods

Mexico$set_region_codes()

Mexico$download()

Mexico$clean_common()

Mexico$clean_level_1()

Mexico$clean_level_2()

Mexico$clone()

Inherited methods

covidregionaldata::DataClass$available_regions()

covidregionaldata::DataClass$clean()

covidregionaldata::DataClass$download_JSON()

covidregionaldata::DataClass$filter()

covidregionaldata::DataClass$get()

covidregionaldata::DataClass$initialize()

covidregionaldata::DataClass$process()

covidregionaldata::DataClass$return()

covidregionaldata::DataClass$summary()

covidregionaldata::DataClass$test()

Method set_region_codes()

Set up a table of region codes for clean data

Usage

Mexico$set_region_codes()

Method download()

Data download() function for Mexico data. This replaces the generic download function in DataClass(). To get the latest data use a PHP script from the website.

Usage

Mexico$download()

Method clean_common()

Common Data Cleaning

Usage

Mexico$clean_common()

Method clean_level_1()

Estados Level Data Cleaning

Usage

Mexico$clean_level_1()

Method clean_level_2()

Municipality Level Data Cleaning

Usage

Mexico$clean_level_2()

Method clone()

The objects of this class are cloneable with this method.

Usage

Mexico$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

if (FALSE) {
region <- Mexico$new(verbose = TRUE, steps = TRUE, get = TRUE)
region$return()
}