Custom CSV reading function

Checks for use of memoise and then uses vroom::vroom.

csv_reader(file, verbose = FALSE, guess_max = 1000, ...)

Arguments

file
A URL or filepath to a CSV

verbose
Logical, defaults to TRUE. Should verbose processing messages and warnings be returned.

guess_max
Maximum number of records to use for guessing column types. Defaults to a 1000.

...
extra parameters to be passed to vroom::vroom

Value

A data table