Get patient line list data

Provides a public international patient line list from January 2020 to June 2020.

This version of the line list has stopped updating. The new version of the line list is free but requires a login.

See: https://global.health/

get_linelist(clean = TRUE, report_delay_only = FALSE)

Arguments

clean
Logical, defaults to TRUE. Should the data returned be cleaned for use.

report_delay_only
Logical, defaults to FALSE. Should only certain variables (id, country, onset date, days' delay), and observations (patients with a report delay) be returned

Source

https://github.com/beoutbreakprepared/nCoV2019

Value

A line list of reported cases of COVID-19

Examples

if (FALSE) {
# Get the complete linelist
get_linelist()

# Return the report delay only
get_linelist(report_delay_only = TRUE)
}