Internal Shared Regional Dataset Processing

Internal shared regional data cleaning designed to be called by process.

process_internal(
  clean_data,
  level,
  group_vars,
  totals = FALSE,
  localise = TRUE,
  verbose = TRUE,
  process_fns
)

Arguments

clean_data
The clean data for a class, e.g. Italy$data$clean

level
The level of the data, e.g. 'level_1_region'

group_vars
Grouping variables, used to for grouping and to localise names. It is assumed that the first entry indicates the main region variable and the second indicates the geocode for this variable.

totals
Logical, defaults to FALSE. If `TRUE``, returns totalled data per region up to today's date. If FALSE, returns the full dataset stratified by date and region.

localise
Logical, defaults to TRUE. Should region names be localised.

verbose
Logical, defaults to TRUE. Should verbose processing messages and warnings be returned.

process_fns
array, additional functions to be called after default processing steps

See also