Lithuania Class for downloading, cleaning and processing notification data

Information for downloading, cleaning and processing COVID-19 region level 1 and 2 data for Lithuania.

Source

https://hub.arcgis.com/datasets/d49a63c934be4f65a93b6273785a8449_0

OSP Data fields

The Official Statistics Portal (OSP) provides many data series in their table.

The full range of these vectors can be returned by setting all_osp_fields to TRUE.

The following describes the data provided by the OSP.

field	description

date	the reporting day during which the events occurred or at the end of which the accounting was performed

municipality_code *	code of the municipality assigned to persons

municipality_name +	the name of the municipality assigned to the persons

population	population size according to the data of the beginning of 2021, according to the declared place of residence

ab_pos_day	Number of positive antibody test responses, days

ab_neg_day	Number of negative antibody test responses, days

ab_tot_day	Number of antibody tests, daily

ab_prc_day	Percentage of positive antibody test responses per day

ag_pos_day	Number of positive antigen test responses, daily

ag_neg_day	Number of negative antigen test responses, daily

ag_tot_day	Number of antigen tests, daily

ag_prc_day	Percentage of positive responses to antigen tests per day

pcr_pos_day	number of positive PCR test responses, daily

pcr_neg_day	Number of PCR test negative responses, daily

pcr_tot_day	number of PCR tests per day

pcr_prc_day	Percentage of positive PCR test responses per day

dgn_pos_day	Number of positive answers to diagnostic tests / tests, days

dgn_neg_day	Number of negative answers to diagnostic tests / tests, days

dgn_prc_day	Number of diagnostic examinations / tests, days

dgn_tot_day	Percentage of positive answers to diagnostic tests / tests per day

dgn_tot_day_gmp	Number of diagnostic examinations / tests of samples collected at mobile points, days

daily_deaths_def1	The number of new deaths per day according to the (narrowest) COVID death definition No. 1. #

daily_deaths_def2	Number of new deaths per day according to COVID death definition No. 2. #

daily_deaths_def3	Number of new deaths per day according to COVID death definition No. 3. #

daily_deaths_all	Daily deaths in Lithuania (by date of death)

incidence +	Number of new COVID cases per day (laboratory or physician confirmed)

cumulative_totals +	Total number of COVID cases (laboratory or physician confirmed)

active_de_jure	Declared number of people with COVID

active_sttstcl	Statistical number of people with COVID

dead_cases	The number of dead persons who were ever diagnosed with COVID

recovered_de_jure	Declared number of recovered live persons

recovered_sttstcl	Statistical number of recovered live persons

map_colors $	The map colour-coding for the municipality, based on averages of test positivity and incidence per capita

* The municipality_code is discarded since it does not correspond to ISO-3166:2 codes used elsewhere in the package.

+ These fields are renamed but returned unmodified.

# Lithuania offers counts according to three different definitions of whether a death is attributable to COVID-19.

$ This field is not recalculated for counties and is deleted.

Criteria for attributing deaths

Beginning in February 2021 the OSP publishes death counts according to three different criteria, from most to least strictly attributed to COVID-19.

of Number of deaths with COVID-19 (coronavirus infection) as the leading cause of death. The indicator is calculated by summing all registered records of medical form E106 (unique persons), in which the main cause of death is IPC disease codes U07.1 or U07.2. Deaths due to external causes are not included (ICD disease codes are V00-Y36, or Y85-Y87, or Y89, or S00-T79, or T89-T98).

with Number of deaths with COVID-19 (coronavirus infection) of any cause of death. The indicator is calculated by summing all registered records of the medical form E106 (unique persons), in which the ICD disease codes U07.1, U07.2, U07.3, U07.4, U07.5 are indicated as the main, direct, intermediate cause of death or other important pathological condition, or identified as related to COVID-19 disease (coronavirus infection). Deaths due to external causes are not included (ICD disease codes are V00-Y36, or Y85-Y87, or Y89, or S00-T79, or T89-T98).

after Number of deaths from any cause of COVID-19 or COVID-19 deaths due to non-external causes within 28 days. The indicator is calculated by summing all registered records of the medical form E106 (unique persons), in which the ICD disease codes U07.1, U07.2, U07.3, U07.4, U07 are indicated as the main, direct, intermediate cause of death or other important pathological condition, or identified as related to COVID-19 disease (coronavirus infection) and all records of medical form E106 (unique individuals) where the person died within the last 28 days after receiving a positive diagnostic response to the SARS-CoV-2 test or had an entry in medical form E025 with ICD disease code U07.2 or U07.1. Deaths due to external causes are not included (ICD disease codes are V00-Y36, or Y85-Y87, or Y89, or S00-T79, or T89-T98).

The number of deaths reported in the last day is preliminary and increases by about 20-40% in a few days. Such a "delay" in the data is natural: for example, for those who died last night, a death certificate is likely to be issued as soon as this report is published this morning.

De jure and statistical counts

Beginning in February 2021 the OSP makes statistical estimates of the number of recovered and active cases, since review of the data showed that some cases individuals still considered as active cases had recovered, but not documented or registered as such.

These are listed as by the OSP as active_de_jure and recovered_de_jure (officially still considered sick), and active_sttstcl and recovered_sttstcl (an estimate of how many of these are still ill).

See also

Super class

covidregionaldata::DataClass -> Lithuania

Public fields

origin

name of origin to fetch data for

supported_levels

A list of supported levels.

supported_region_names

A list of region names in order of level.

supported_region_codes

A list of region codes in order of level.

common_data_urls

List of named links to raw data that are common across levels.

source_data_cols

existing columns within the raw data

death_definition

which criteria of deaths attributed to COVID to use

recovered_definition

whether to use the official counts of recovered cases or the statistical estimates provided by OSP

all_osp_fields

whether to return all the data vectors provided by OSP

national_data

whether to return data rows for national results

Methods

Public methods

Lithuania$set_region_codes()

Lithuania$clean_common()

Lithuania$clean_level_1()

Lithuania$new()

Lithuania$clone()

Inherited methods

covidregionaldata::DataClass$available_regions()

covidregionaldata::DataClass$clean()

covidregionaldata::DataClass$download()

covidregionaldata::DataClass$filter()

covidregionaldata::DataClass$get()

covidregionaldata::DataClass$process()

covidregionaldata::DataClass$return()

covidregionaldata::DataClass$summary()

covidregionaldata::DataClass$test()

Method set_region_codes()

Set up a table of region codes for clean data

Usage

Lithuania$set_region_codes()

Method set_region_codes()

Set up a table of region codes for clean data

Usage

Lithuania$set_region_codes()

Method clean_common()

Common data cleaning for both levels

Usage

Lithuania$clean_common()

Method clean_common()

Common data cleaning for both levels

Usage

Lithuania$clean_common()

Method clean_level_1()

Lithuania Specific County Level Data Cleaning

Aggregates data to the level 1 (county) regional level. Data is provided by the source at the level 2 (municipality) regional level.

Usage

Lithuania$clean_level_1()

Method clean_level_1()

Lithuania Specific County Level Data Cleaning

Aggregates data to the level 1 (county) regional level. Data is provided by the source at the level 2 (municipality) regional level.

Usage

Lithuania$clean_level_1()

Method new()

Initialize the country

Usage

Lithuania$new(
  death_definition = "of",
  recovered_definition = "official",
  all_osp_fields = FALSE,
  national_data = FALSE,
  ...
)

Arguments

death_definition

A character string. Determines which criteria for attributing deaths to COVID is used. Should be "of", "with", or "after". Can also be "daily_deaths_def1", "daily_deaths_def2", or "daily_deaths_def3". (Defaults to "of", the strictest definition.)

recovered_definition

A character string. Determines whether the count of officially-recovered (de jure) cases is used, or the statistical estimate provided by OSP. Should be "official" or "statistical". (Defaults to "official".)

all_osp_fields

A logical scalar. Should all the meaningful data fields from the OSP source be returned? (Defaults FALSE)

national_data

A logical scalar. Should national values be returned? (Defaults FALSE)

...

Parameters passed to DataClass() initalize

Method new()

Initialize the country

Usage

Lithuania$new(
  death_definition = "of",
  recovered_definition = "official",
  all_osp_fields = FALSE,
  national_data = FALSE,
  ...
)

Arguments

death_definition

A character string. Determines which criteria for attributing deaths to COVID is used. Should be "of", "with", or "after". Can also be "daily_deaths_def1", "daily_deaths_def2", or "daily_deaths_def3". (Defaults to "of", the strictest definition.)

recovered_definition

A character string. Determines whether the count of officially-recovered (de jure) cases is used, or the statistical estimate provided by OSP. Should be "official" or "statistical". (Defaults to "official".)

all_osp_fields

A logical scalar. Should all the meaningful data fields from the OSP source be returned? (Defaults FALSE)

national_data

A logical scalar. Should national values be returned? (Defaults FALSE)

...

Parameters passed to DataClass() initalize

Method clone()

The objects of this class are cloneable with this method.

Usage

Lithuania$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Method clone()

The objects of this class are cloneable with this method.

Usage

Lithuania$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

if (FALSE) {
region <- Lithuania$new(verbose = TRUE, steps = TRUE, get = TRUE)
}