Get the countries data is available for

Show what countries have what level data available. The function searches the environment for R6 class objects and extracts the summary information from the available classes using their summary methods.

get_available_datasets(type)

Arguments

type
A character vector indicating the types of data to return. Current options include "national" (which are datasets at the national level which inherit from CountryDataClass) and "regional" (which are datasets at the regional level which inherit directly from DataClass).

Value

A list of available countries and the region level data is available for

Examples

# see all available datasets
get_available_datasets()

#> # A tibble: 18 x 8
#>    origin   class level_1_region level_2_region level_3_region type  data_urls  
#>    <chr>    <chr> <chr>          <chr>          <chr>          <chr> <chr>      
#>  1 Belgium  Belg… region         province       NA             regi… https://ep…
#>  2 Brazil   Braz… state          city           NA             regi… https://gi…
#>  3 Canada   Cana… province       NA             NA             regi… https://he…
#>  4 Colombia Colo… departamento   NA             NA             regi… https://ra…
#>  5 Cuba     Cuba  provincia      NA             NA             regi… https://co…
#>  6 Europea… ECDC  country        NA             NA             nati… https://op…
#>  7 France   Fran… region         department     NA             regi… https://ww…
#>  8 Germany  Germ… bundesland     landkreis      NA             regi… https://op…
#>  9 Google   Goog… country        subregion      subregion2     nati… https://st…
#> 10 India    India state          NA             NA             regi… https://ap…
#> 11 Italy    Italy regioni        NA             NA             regi… https://ra…
#> 12 John Ho… JHU   country        region         NA             nati… https://ra…
#> 13 Lithuan… Lith… county         municipality   NA             regi… https://op…
#> 14 Mexico   Mexi… estado         municipio      NA             regi… Downloads/…
#> 15 South A… Sout… province       NA             NA             regi… https://ra…
#> 16 United … UK    region         authority      NA             regi… https://ww…
#> 17 United … USA   state          county         NA             regi… https://ra…
#> 18 World H… WHO   country        NA             NA             nati… https://co…
#> # … with 1 more variable: source_data_cols <chr>

# see only national level datasets
get_available_datasets("national")

#> # A tibble: 4 x 8
#>   origin   class level_1_region level_2_region level_3_region type  data_urls   
#>   <chr>    <chr> <chr>          <chr>          <chr>          <chr> <chr>       
#> 1 Europea… ECDC  country        NA             NA             nati… https://ope…
#> 2 Google   Goog… country        subregion      subregion2     nati… https://sto…
#> 3 John Ho… JHU   country        region         NA             nati… https://raw…
#> 4 World H… WHO   country        NA             NA             nati… https://cov…
#> # … with 1 more variable: source_data_cols <chr>

# see only regional level datasets
get_available_datasets("regional")

#> # A tibble: 14 x 8
#>    origin  class  level_1_region level_2_region level_3_region type  data_urls  
#>    <chr>   <chr>  <chr>          <chr>          <chr>          <chr> <chr>      
#>  1 Belgium Belgi… region         province       NA             regi… https://ep…
#>  2 Brazil  Brazil state          city           NA             regi… https://gi…
#>  3 Canada  Canada province       NA             NA             regi… https://he…
#>  4 Colomb… Colom… departamento   NA             NA             regi… https://ra…
#>  5 Cuba    Cuba   provincia      NA             NA             regi… https://co…
#>  6 France  France region         department     NA             regi… https://ww…
#>  7 Germany Germa… bundesland     landkreis      NA             regi… https://op…
#>  8 India   India  state          NA             NA             regi… https://ap…
#>  9 Italy   Italy  regioni        NA             NA             regi… https://ra…
#> 10 Lithua… Lithu… county         municipality   NA             regi… https://op…
#> 11 Mexico  Mexico estado         municipio      NA             regi… Downloads/…
#> 12 South … South… province       NA             NA             regi… https://ra…
#> 13 United… UK     region         authority      NA             regi… https://ww…
#> 14 United… USA    state          county         NA             regi… https://ra…
#> # … with 1 more variable: source_data_cols <chr>