Initialise a child class of DataClass if it exists

This function initialises classes based on the DataClass() which allows documented downloading, cleaning, and processing. See the examples for some potential use cases and the DataClass() documentation for more details.

initialise_dataclass(
  class = character(),
  level = "1",
  totals = FALSE,
  localise = TRUE,
  regions,
  verbose = TRUE,
  steps = FALSE,
  get = FALSE,
  type = c("national", "regional"),
  ...
)

Arguments

class
A character string specifying the DataClass() to initialise. Not case dependent and matching is based on either the class name or the its country definition. For a list of options use get_available_datasets().

level
A character string indicating the target administrative level of the data with the default being "1". Currently supported options are level 1 ("1) and level 2 ("2"). Use get_available_datasets() for supported options by dataset.

totals
Logical, defaults to FALSE. If TRUE, returns totalled data per region up to today's date. If FALSE, returns the full dataset stratified by date and region.

localise
Logical, defaults to TRUE. Should region names be localised.

regions
A character vector of target regions to be assigned to the target_regions field and used to filter the returned data.

verbose
Logical, defaults to TRUE. Should verbose processing messages and warnings be returned.

steps
Logical, defaults to FALSE. Should all processing and cleaning steps be kept and output in a list.

get
Logical, defaults to FALSE. Should the class get method be called (this will download, clean, and process data at initialisation).

type
A character vector indicating the types of data to return. Current options include "national" (which are datasets at the national level which inherit from CountryDataClass) and "regional" (which are datasets at the regional level which inherit directly from DataClass()).

...
Additional arguments to pass to class specific functionality.

Value

An initialised version of the target class if available, e.g. Italy()

See also

Examples

# set up a cache to store data to avoid downloading repeatedly
start_using_memoise()
#> Using a cache at: /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//RtmpDO1S4j

# check currently available datasets
get_available_datasets()
#> # A tibble: 18 x 8
#>    origin   class level_1_region level_2_region level_3_region type  data_urls  
#>    <chr>    <chr> <chr>          <chr>          <chr>          <chr> <chr>      
#>  1 Belgium  Belg… region         province       NA             regi… https://ep…
#>  2 Brazil   Braz… state          city           NA             regi… https://gi…
#>  3 Canada   Cana… province       NA             NA             regi… https://he…
#>  4 Colombia Colo… departamento   NA             NA             regi… https://ra…
#>  5 Cuba     Cuba  provincia      NA             NA             regi… https://co…
#>  6 Europea… ECDC  country        NA             NA             nati… https://op…
#>  7 France   Fran… region         department     NA             regi… https://ww…
#>  8 Germany  Germ… bundesland     landkreis      NA             regi… https://op…
#>  9 Google   Goog… country        subregion      subregion2     nati… https://st…
#> 10 India    India state          NA             NA             regi… https://ap…
#> 11 Italy    Italy regioni        NA             NA             regi… https://ra…
#> 12 John Ho… JHU   country        region         NA             nati… https://ra…
#> 13 Lithuan… Lith… county         municipality   NA             regi… https://op…
#> 14 Mexico   Mexi… estado         municipio      NA             regi… Downloads/…
#> 15 South A… Sout… province       NA             NA             regi… https://ra…
#> 16 United … UK    region         authority      NA             regi… https://ww…
#> 17 United … USA   state          county         NA             regi… https://ra…
#> 18 World H… WHO   country        NA             NA             nati… https://co…
#> # … with 1 more variable: source_data_cols <chr>

# initialise a data set in the United Kingdom
# at the UTLA level
utla <- UK$new(level = "2")

# download UTLA data
utla$download()
#> Downloading UK data.
#> Downloading data from https://api.coronavirus.data.gov.uk/v2/data?areaType=utla&metric=newCasesBySpecimenDate&metric=cumCasesBySpecimenDate&metric=newCasesByPublishDate&metric=cumCasesByPublishDate&format=csv
#> Rows: 91,426
#> Columns: 8
#> Delimiter: ","
#> chr  [3]: areaCode, areaName, areaType
#> dbl  [4]: cumCasesByPublishDate, cumCasesBySpecimenDate, newCasesByPublishDate, newCases...
#> date [1]: date
#> 
#> Use `spec()` to retrieve the guessed column specification
#> Pass a specification to the `col_types` argument to quiet this message
#> Downloading data from https://api.coronavirus.data.gov.uk/v2/data?areaType=utla&metric=newDeaths28DaysByPublishDate&metric=cumDeaths28DaysByPublishDate&metric=newDeaths28DaysByDeathDate&metric=cumDeaths28DaysByDeathDate&format=csv
#> Rows: 79,297
#> Columns: 8
#> Delimiter: ","
#> chr  [3]: areaCode, areaName, areaType
#> dbl  [4]: cumDeaths28DaysByDeathDate, cumDeaths28DaysByPublishDate, newDeaths28DaysByDea...
#> date [1]: date
#> 
#> Use `spec()` to retrieve the guessed column specification
#> Pass a specification to the `col_types` argument to quiet this message
#> Downloading data from https://api.coronavirus.data.gov.uk/v2/data?areaType=utla&metric=newTestsByPublishDate&metric=cumTestsByPublishDate&metric=newAdmissions&metric=cumAdmissions&format=csv
#> Downloading data from https://api.coronavirus.data.gov.uk/v2/data?areaType=utla&metric=newPillarOneTestsByPublishDate&metric=newPillarTwoTestsByPublishDate&metric=newPillarThreeTestsByPublishDate&metric=newPillarFourTestsByPublishDate&format=csv

# clean UTLA data
utla$clean()
#> Cleaning data

# inspect available level 1 regions
utla$available_regions(level = "1")
#>  [1] "North East"               "Yorkshire and The Humber"
#>  [3] "North West"               "West Midlands"           
#>  [5] "London"                   "Scotland"                
#>  [7] "Wales"                    "South West"              
#>  [9] "East of England"          "East Midlands"           
#> [11] "South East"               "Northern Ireland"        

# filter data to the East of England
utla$filter("East of England")
#> Filtering data to: East of England

# process UTLA data
utla$process()
#> Processing data

# return processed and filtered data
utla$return()
#> # A tibble: 4,796 x 24
#>    date       region      region_code authority      local_authority_… cases_new
#>    <date>     <chr>       <chr>       <chr>          <chr>                 <dbl>
#>  1 2020-02-26 East of En… E12000006   Bedford        E06000055                NA
#>  2 2020-02-26 East of En… E12000006   Cambridgeshire E10000003                NA
#>  3 2020-02-26 East of En… E12000006   Central Bedfo… E06000056                NA
#>  4 2020-02-26 East of En… E12000006   Essex          E10000012                NA
#>  5 2020-02-26 East of En… E12000006   Hertfordshire  E10000015                 1
#>  6 2020-02-26 East of En… E12000006   Luton          E06000032                NA
#>  7 2020-02-26 East of En… E12000006   Norfolk        E10000020                NA
#>  8 2020-02-26 East of En… E12000006   Peterborough   E06000031                NA
#>  9 2020-02-26 East of En… E12000006   Southend-on-S… E06000033                NA
#> 10 2020-02-26 East of En… E12000006   Suffolk        E10000029                NA
#> # … with 4,786 more rows, and 18 more variables: cases_total <dbl>,
#> #   deaths_new <dbl>, deaths_total <dbl>, recovered_new <dbl>,
#> #   recovered_total <dbl>, hosp_new <dbl>, hosp_total <dbl>, tested_new <dbl>,
#> #   tested_total <dbl>, areaType <chr>, cumCasesByPublishDate <dbl>,
#> #   cumCasesBySpecimenDate <dbl>, newCasesByPublishDate <dbl>,
#> #   newCasesBySpecimenDate <dbl>, cumDeaths28DaysByDeathDate <dbl>,
#> #   cumDeaths28DaysByPublishDate <dbl>, newDeaths28DaysByDeathDate <dbl>,
#> #   newDeaths28DaysByPublishDate <dbl>

# inspect all data steps
utla$data
#> $raw
#> $raw$utla
#> # A tibble: 91,426 x 12
#>    areaCode  areaName    areaType date       cumCasesByPublis… cumCasesBySpecim…
#>    <chr>     <chr>       <chr>    <date>                 <dbl>             <dbl>
#>  1 E06000003 Redcar and… utla     2021-05-06              9341                NA
#>  2 E06000014 York        utla     2021-05-06             12297                NA
#>  3 E06000050 Cheshire W… utla     2021-05-06             21768                NA
#>  4 E08000001 Bolton      utla     2021-05-06             26517                NA
#>  5 E08000016 Barnsley    utla     2021-05-06             19607                NA
#>  6 E08000031 Wolverhamp… utla     2021-05-06             24864                NA
#>  7 E08000032 Bradford    utla     2021-05-06             49709                NA
#>  8 E09000018 Hounslow    utla     2021-05-06             25008                NA
#>  9 E09000032 Wandsworth  utla     2021-05-06             22269                NA
#> 10 E09000033 Westminster utla     2021-05-06             14276                NA
#> # … with 91,416 more rows, and 6 more variables: newCasesByPublishDate <dbl>,
#> #   newCasesBySpecimenDate <dbl>, cumDeaths28DaysByDeathDate <dbl>,
#> #   cumDeaths28DaysByPublishDate <dbl>, newDeaths28DaysByDeathDate <dbl>,
#> #   newDeaths28DaysByPublishDate <dbl>
#> 
#> 
#> $clean
#> # A tibble: 91,426 x 18
#>    level_2_region_co… level_2_region      areaType date       cumCasesByPublish…
#>    <chr>              <chr>               <chr>    <date>                  <dbl>
#>  1 E06000003          Redcar and Clevela… utla     2021-05-06               9341
#>  2 E06000014          York                utla     2021-05-06              12297
#>  3 E06000050          Cheshire West and … utla     2021-05-06              21768
#>  4 E08000001          Bolton              utla     2021-05-06              26517
#>  5 E08000016          Barnsley            utla     2021-05-06              19607
#>  6 E08000031          Wolverhampton       utla     2021-05-06              24864
#>  7 E08000032          Bradford            utla     2021-05-06              49709
#>  8 E09000018          Hounslow            utla     2021-05-06              25008
#>  9 E09000032          Wandsworth          utla     2021-05-06              22269
#> 10 E09000033          Westminster         utla     2021-05-06              14276
#> # … with 91,416 more rows, and 13 more variables: cumCasesBySpecimenDate <dbl>,
#> #   newCasesByPublishDate <dbl>, newCasesBySpecimenDate <dbl>,
#> #   cumDeaths28DaysByDeathDate <dbl>, cumDeaths28DaysByPublishDate <dbl>,
#> #   newDeaths28DaysByDeathDate <dbl>, newDeaths28DaysByPublishDate <dbl>,
#> #   cases_new <dbl>, cases_total <dbl>, deaths_new <dbl>, deaths_total <dbl>,
#> #   level_1_region_code <chr>, level_1_region <chr>
#> 
#> $filtered
#> # A tibble: 4,692 x 18
#>    level_2_region_code level_2_region areaType date       cumCasesByPublishDate
#>    <chr>               <chr>          <chr>    <date>                     <dbl>
#>  1 E10000015           Hertfordshire  utla     2021-05-06                 77033
#>  2 E10000015           Hertfordshire  utla     2021-05-05                    NA
#>  3 E10000015           Hertfordshire  utla     2021-05-04                    NA
#>  4 E10000015           Hertfordshire  utla     2021-05-03                    NA
#>  5 E10000015           Hertfordshire  utla     2021-05-02                    NA
#>  6 E10000015           Hertfordshire  utla     2021-05-01                    NA
#>  7 E10000015           Hertfordshire  utla     2021-04-30                    NA
#>  8 E10000015           Hertfordshire  utla     2021-04-29                    NA
#>  9 E10000015           Hertfordshire  utla     2021-04-28                    NA
#> 10 E10000015           Hertfordshire  utla     2021-04-27                    NA
#> # … with 4,682 more rows, and 13 more variables: cumCasesBySpecimenDate <dbl>,
#> #   newCasesByPublishDate <dbl>, newCasesBySpecimenDate <dbl>,
#> #   cumDeaths28DaysByDeathDate <dbl>, cumDeaths28DaysByPublishDate <dbl>,
#> #   newDeaths28DaysByDeathDate <dbl>, newDeaths28DaysByPublishDate <dbl>,
#> #   cases_new <dbl>, cases_total <dbl>, deaths_new <dbl>, deaths_total <dbl>,
#> #   level_1_region_code <chr>, level_1_region <chr>
#> 
#> $processed
#> # A tibble: 4,796 x 24
#>    date       region      region_code authority      local_authority_… cases_new
#>    <date>     <chr>       <chr>       <chr>          <chr>                 <dbl>
#>  1 2020-02-26 East of En… E12000006   Bedford        E06000055                NA
#>  2 2020-02-26 East of En… E12000006   Cambridgeshire E10000003                NA
#>  3 2020-02-26 East of En… E12000006   Central Bedfo… E06000056                NA
#>  4 2020-02-26 East of En… E12000006   Essex          E10000012                NA
#>  5 2020-02-26 East of En… E12000006   Hertfordshire  E10000015                 1
#>  6 2020-02-26 East of En… E12000006   Luton          E06000032                NA
#>  7 2020-02-26 East of En… E12000006   Norfolk        E10000020                NA
#>  8 2020-02-26 East of En… E12000006   Peterborough   E06000031                NA
#>  9 2020-02-26 East of En… E12000006   Southend-on-S… E06000033                NA
#> 10 2020-02-26 East of En… E12000006   Suffolk        E10000029                NA
#> # … with 4,786 more rows, and 18 more variables: cases_total <dbl>,
#> #   deaths_new <dbl>, deaths_total <dbl>, recovered_new <dbl>,
#> #   recovered_total <dbl>, hosp_new <dbl>, hosp_total <dbl>, tested_new <dbl>,
#> #   tested_total <dbl>, areaType <chr>, cumCasesByPublishDate <dbl>,
#> #   cumCasesBySpecimenDate <dbl>, newCasesByPublishDate <dbl>,
#> #   newCasesBySpecimenDate <dbl>, cumDeaths28DaysByDeathDate <dbl>,
#> #   cumDeaths28DaysByPublishDate <dbl>, newDeaths28DaysByDeathDate <dbl>,
#> #   newDeaths28DaysByPublishDate <dbl>
#> 
#> $return
#> [1] NA
#> 

# initialise Italian data, download, clean and process it
italy <- initialise_dataclass("Italy", get = TRUE)
#> Downloading data from https://raw.githubusercontent.com/pcm-dpc/COVID-19/master/dati-regioni/dpc-covid19-ita-regioni.csv
#> Rows: 9,198
#> Columns: 30
#> Delimiter: ","
#> chr  [ 7]: stato, denominazione_regione, note, note_test, note_casi, codice_nuts_1, codice...
#> dbl  [22]: codice_regione, lat, long, ricoverati_con_sintomi, terapia_intensiva, totale_os...
#> dttm [ 1]: data
#> 
#> Use `spec()` to retrieve the guessed column specification
#> Pass a specification to the `col_types` argument to quiet this message
#> Warning: One or more parsing issues, see `problems()` for details
#> Cleaning data
#> Processing data
italy$return()
#> # A tibble: 8,760 x 13
#>    date       regioni   iso_3166_2 cases_new cases_total deaths_new deaths_total
#>    <date>     <chr>     <chr>          <dbl>       <dbl>      <dbl>        <dbl>
#>  1 2020-02-24 Abruzzo   IT-65              0           0          0            0
#>  2 2020-02-24 Basilica… IT-77              0           0          0            0
#>  3 2020-02-24 Calabria  IT-78              0           0          0            0
#>  4 2020-02-24 Campania  IT-72              0           0          0            0
#>  5 2020-02-24 Emilia-R… IT-45             18          18          0            0
#>  6 2020-02-24 Friuli V… IT-36              0           0          0            0
#>  7 2020-02-24 Lazio     IT-62              3           3          0            0
#>  8 2020-02-24 Liguria   IT-42              0           0          0            0
#>  9 2020-02-24 Lombardia IT-25            172         172          6            6
#> 10 2020-02-24 Marche    IT-57              0           0          0            0
#> # … with 8,750 more rows, and 6 more variables: recovered_new <dbl>,
#> #   recovered_total <dbl>, hosp_new <dbl>, hosp_total <dbl>, tested_new <dbl>,
#> #   tested_total <dbl>

# initialise ECDC data, fully process it, and return totals
ecdc <- initialise_dataclass("ecdc", get = TRUE, totals = TRUE)
#> Downloading data from https://opendata.ecdc.europa.eu/covid19/casedistribution/csv
#> Rows: 61,900
#> Columns: 12
#> Delimiter: ","
#> chr [5]: dateRep, countriesAndTerritories, geoId, countryterritoryCode, continentExp
#> dbl [7]: day, month, year, cases, deaths, popData2019, Cumulative_number_for_14_days_of_...
#> 
#> Use `spec()` to retrieve the guessed column specification
#> Pass a specification to the `col_types` argument to quiet this message
#> Cleaning data
#> Processing data
ecdc$return()
#> # A tibble: 213 x 7
#>    country        iso_code cases_total deaths_total recovered_total hosp_total
#>    <chr>          <chr>          <dbl>        <dbl>           <dbl>      <dbl>
#>  1 United States  US          16256754       299177               0          0
#>  2 India          IN           9884100       143355               0          0
#>  3 Brazil         BR           6901952       181402               0          0
#>  4 Russia         RU           2653928        46941               0          0
#>  5 France         FR           2377618        57911               0          0
#>  6 United Kingdom UK           1849403        64170               0          0
#>  7 Italy          IT           1843860        64551               0          0
#>  8 Spain          ES           1731660        49544               0          0
#>  9 Argentina      AR           1498160        40766               0          0
#> 10 Colombia       CO           1425774        39053               0          0
#> # … with 203 more rows, and 1 more variable: tested_total <dbl>